apps/auth/pyproject.toml
[project]
name = "mint-auth"
version = "2.0.0"
description = "auth service for mint"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "fastapi[all]>=0.135.1",
    "joserfc>=1.6.3",
    "pydantic-settings==2.12.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.48",
    "mint-shared",
    "asyncpg>=0.31.0",
    "faststream[kafka]>=0.6.7",
    "sreekarnv-fastauth[email]>=0.5.3",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0",
    "opentelemetry-instrumentation-httpx>=0.46b0",
    "opentelemetry-exporter-otlp-proto-http>=1.25.0",
]

[project.scripts]
dev = "auth.main:start"
makemigrations = "cli:makemigrations"
migrate       = "cli:migrate"
downgrade     = "cli:downgrade"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auth"]
sources  = ["src"]

[tool.uv.sources]
mint-shared = { workspace = true }
